Доработка сервиса магазина(заказ, доставка, оплата).

+++ commerce/interaction-api/src/main/java/ru/yandex/practicum/commerce/interactionapi/warehouse/DimensionDto.java
@@ -5,6 +5,8 @@ import jakarta.validation.constraints.NotNull;
 import lombok.*;
 import lombok.experimental.FieldDefaults;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @AllArgsConstructor
@@ -14,11 +16,11 @@ public class DimensionDto {
 
     @NotNull(message = "Необходима ширина товара.")
     @Min(value = 1, message = "Ширина должна быть минимум 1.")
-    Double width;
+    BigDecimal width;
     @NotNull(message = "Необходима высота товара.")
     @Min(value = 1, message = "Высота должна быть минимум 1.")
-    Double height;
+    BigDecimal height;
     @NotNull(message = "Необходима глубина товара.")
     @Min(value = 1, message = "Глубина должна быть минимум 1.")
-    Double depth;
+    BigDecimal depth;
 }
